requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
numpy==1.24.4
realesrgan-ncnn-py==1.0.2
python-multipart==0.0.6